requirements.txt
asgiref==3.9.1
attrs==25.3.0
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.24.1
jsonschema-specifications==2025.4.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.26.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
requests==2.32.3
